REEB-876 PointGoods.pointPrice 改成 long
git-svn-id: http://svn.uhuila.cn/reeb/trunk@2561 96626eaa-6fa6-4472-8f5b-47c901fe3adc

+++ module/sales/app/models/sales/PointGoods.java
@@ -107,7 +107,7 @@ public class PointGoods extends Model {
     @Max(999999)
     @Money
     @Column(name = "point_price")
-    public BigDecimal pointPrice;
+    public Long pointPrice;
 
     //  ======  价格列表结束 ==========
